+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2024-11-23
+
+### Added
+- [audio/.gitignore](./audio/.gitignore)
+- [audio/disc_check.py](./audio/disc_check.py)
+- [audio/mount.sh](./audio/mount.sh)
+- [audio/pico-mount-fix.sh](./audio/pico-mount-fix.sh)
+- [audio/pico_mount_repair.py](./audio/pico_mount_repair.py)
+- [audio/pico_setup.py](./audio/pico_setup.py)
+- [audio/src/code.py](./audio/src/code.py)
+- [audio/src/settings.toml](./audio/src/settings.toml)
+- [start.sh](./start.sh)
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [0.0.1] - 2024-11-23
 
 ### Added